full support for adding #lang lines as kernels
Currently this only works for #langs that have top-interaction
support.

The primary issue here is that the Jupyter cell model means that we
live in the "hopeless" top level all the time. It may very well be
possible to have cells treated as modules that automatically invoke
(provide (all-defined-out)) and to automatically import those modules
into all future cells, but that would be a non-trivial change.